Part 4 - External Clients
This adds support for letting more kinds of clients connect and without
being created by the WM itself. This should open up the ability to use
the normal arcan native clients, that is aloadimage for image loading,
afsrv_game for libretro cores, afsrv_decode for video playback and
Xarcan for running an Xserver.

Client 'adoption' is now also supported, which permits other WMs to
switch to this one, or it being used as a 'fallback' on a scripting
error, allowing script-level crash recovery.